factor ArrayUUIDsField validation out
